+++ CHANGES.md
@@ -16,10 +16,12 @@
     * `augur curate rename --field-map`
     * `augur curate transform-strain-name --backup-fields`
 * `augur curate format-dates --expected-date-formats` help text has been improved with clarifications regarding how values provided interact with builtin formats and how to match masked date parts. [#1707][], [#1718][] (@victorlin)
+* parse: Transform strain names the same way in both metadata and sequences instead of only transforming sequences. [#1712][] (@huddlej)
 
 [#1688]: https://github.com/nextstrain/augur/pull/1688
 [#1690]: https://github.com/nextstrain/augur/pull/1690
 [#1707]: https://github.com/nextstrain/augur/issues/1707
+[#1712]: https://github.com/nextstrain/augur/pull/1712
 [#1715]: https://github.com/nextstrain/augur/pull/1715
 [#1716]: https://github.com/nextstrain/augur/pull/1716
 [#1718]: https://github.com/nextstrain/augur/pull/1718